Built base of server-client applications
The server runs on a port and stores a list of client messages. It will
send back all client messages to each client when a change occurs.

Clients connect to the server and can type any string. Strings from
other clients and yourself can be seen. It will update when anyone
modify the string and send it.