Engine: Prevent too fast frame speed
On some machines, when V-Sync is turned off in settings of video driver but operating system still report that V-Sync is available for use, the game thinks V-Sync is supported and turns it on. When V-Sync is not really on, the game begins to process game as fast as CPU and renderer can. In result, the game speed is going like fast-forwarding of video tape.