Remove the PgeGameSave library
This is the unfinished library which got outdated and it was never used.

That was an idea in the past to implement the "better" gamesave format that will use the database tables as one single file, etc. And there was a sort of ideas like "soft-save" (when game makes automatic saves) and "hard-save" (when player requests the saving of the game manually). Using the "soft-save" as a way to resume the game after the crash, power surge, or sudden killing/closing of the game. So, after rethinking of game saving in whole (for both Moondust Engine and TheXTech), I do keep all gamesaves out of episode directory to keep it being read-only friendly which is critically for Emscripten, 3DS, and other folks, where is no any sort of the file system in a common sense (also three the case when episode got packed into the special archive format, etc.). So, SAVX file is right now can store any user data that can be encoded in any convenient way (for example, JSON) and they were already has the location (global, level-local, world-local) and other cases, just as a SAVX-built-in feature. The SAVX format isn't stress resistant like SQLite, but it's flexible and simple to use at the same time. So, right now I do think to archive this repository and take it again once we REALLY need for it.