Editor: Added the "default" field to supported features list
This is a special field that sets global default value for all features at once. So, if you set the `default = hidden` state, all unlisted fields will be set to the `hidden` automatically. So, this should avlid any unexpected feature introduction at some configuration pacakges.

Should help with #544 in the future.