Editor: Added an ability to change the label font
This update adds an ability to alternate the font that is being used for the floating label in the Editor's scene. However, to change this setup, need to CLOSE the editor, open the settings/pge_editor.ini and find `label-box-font-...` parameters, edit them as you need, and run Editor again. The UI to edit this setup is comming soon.
#413